Reorganize ioc_examples.
