+++ caproto/ioc_examples/hardware_specific/barcode_reader.py
@@ -4,8 +4,9 @@ This example requires a barcode reader and the Python library evdev.
 
 This only works on linux (as it uses the linux kernel input events).
 """
-from caproto.server import pvproperty, PVGroup, template_arg_parser, run
 from evdev import InputDevice, categorize
+
+from caproto.server import PVGroup, pvproperty, run, template_arg_parser
 
 
 class BarcodeIOC(PVGroup):
