+++ caproto/ioc_examples/pathological/reading_counter.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
-from caproto.server import pvproperty, PVGroup, ioc_arg_parser, run
 import collections
+
+from caproto.server import PVGroup, ioc_arg_parser, pvproperty, run
 
 
 class ReadingCounter(PVGroup):
