+++ caproto/ioc_examples/too_clever/dynamic_pvgroups.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-from caproto.server import pvproperty, PVGroup, ioc_arg_parser, run
+from caproto.server import PVGroup, ioc_arg_parser, pvproperty, run
 
 
 class IOCMain(PVGroup):
