+++ caproto/tests/verify_with_catvs.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
-from caproto.server import pvproperty, PVGroup, ioc_arg_parser, run
 from caproto import ChannelType
+from caproto.server import PVGroup, ioc_arg_parser, pvproperty, run
 
 
 class CatvsIOC(PVGroup):
